camera file refactored
