+++ copylot/hardware/orca_camera/demo/camera_demo.py
@@ -1,4 +1,4 @@
-from copylot.hardware.orca_camera.orcacamera import OrcaCamera
+from copylot.hardware.orca_camera.camera import OrcaCamera
 
 if __name__ == '__main__':
     camera = OrcaCamera()
